Mantine UI Integration, Login, Auth Fixes
Installed Mantine UI Library and PostCSS to handle styles in Mantine UIs and avoid conflicts with pre-existing SASS.

Created and initialized `postcss.config.cjs` as per Mantine's requirements

Downgraded Next and NextAuth versions due to bugs in the configuration of auth pages and CSRF.

Enclosed the Providers, Sidebar, and children in a provider given by Mantine.

Changed the existing logo with Bibiliotechai's official and added an image of the UST Blessed Pier Giorgio Frassati Building.

Grouped `login` and `signup` into a route called `(forms).` This would be handled by the `layout.js` in the parent folder.

Modified the `route.js` in the API to override the NextAuth's sign-in page.

The `/login` page is already styled, and the backend side of it is completed.

Used useRef to catch all the real-time inputs in the TextFields. Then, all relevant details are passed onClick with an onSubmit method.